backend/requirements.txt
Django>=5.2,<5.3
django-cors-headers==4.6.0
gunicorn==23.0.0
whitenoise==6.8.2
psycopg[binary]>=3.2,<4.0
dj-database-url==2.3.0
python-dotenv==1.0.1
